refactor: Add 404 handling and redirect functionality
A 404.html file was added for handling unknown routes, and used sessionStorage to store the last known location. In the main index.html file, a script was added to process these redirects, replacing the current URL with the one stored in sessionStorage. The vite.config.ts file was also updated to include the 404.html as part of the rollup options.